move test with their components
