+++ components/Button/Button.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Button from '../components/Button/Button';
+import Button from './Button';
 
 describe('<Button />', () => {
   const testProps = {
